+++ src/app/(auth)/login.tsx
@@ -1,0 +1,1 @@
+export { default } from '@features/auth/screens/Login';
