+++ tsconfig.json
@@ -9,7 +9,8 @@
       "@app/*": ["./src/app/*"],
       "@components/*": ["./src/components/*"],
       "@constants/*": ["./src/constants/*"],
-      "@hooks/*": ["./src/hooks/*"]
+      "@hooks/*": ["./src/hooks/*"],
+      "@features/*": ["./src/features/*"]
     }
   },
   "include": [
